refactor: centralize ScreeningQuestionsRelationManager logic and update references
